requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.23.5
botocore==1.26.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.13
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.0
mysqlclient==2.1.0
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
